pyproject.toml
[project]
name = "booking-engine-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "django>=6.0.1",
  "djangorestframework>=3.16.1",
  "djangorestframework-simplejwt>=5.5.1",
  "pillow>=12.1.0",
  "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
  "django-stubs>=5.2.8",
  "django-types>=0.22.0",
  "djangorestframework-stubs>=3.16.7",
]

[tool.pyright]
reportIncompatibleVariableOverride = false
